+++ src/mqtt.cpp
@@ -57,6 +57,7 @@ void mqtt_publish_door_state()
         case DoorState::Closed: state = MQTT_STATUSDOORCLOSED; break;
         case DoorState::Opening: state = MQTT_STATUSDOOROPENING; break;
         case DoorState::Closing: state = MQTT_STATUSDOORCLOSING; break;
+        case DoorState::Stopped: state = MQTT_STATUSDOORSTOPPED; break;
         case DoorState::Unknown: break;
     }
     if (!mqtt_send(MQTT_TOPICCONTROLGETCURRENTDOORSTATE, state, true)) return;